+++ src/main/java/se2/hanu_hospital/equipment/EquipmentPayload.java
@@ -3,14 +3,14 @@ package se2.hanu_hospital.equipment;
 
 public class EquipmentPayload {
     private String name;
-    private Long price;
+    private Double price;
     private Integer quantity;
 
     public String getName() {
         return name;
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
